+++ build.rs
@@ -48,11 +48,12 @@ fn main() {
     // Add custom functions from Rust
     let mut lib = Module::new();
     combine_with_exported_module!(&mut lib, "rhai_sci_matrix_function", matrix_functions);
-    combine_with_exported_module!(&mut lib, "rhai_sci_utility_functions", util_functions);
+    combine_with_exported_module!(&mut lib, "rhai_sci_miscellaneous_functions", misc_functions);
     combine_with_exported_module!(&mut lib, "rhai_sci_basic_stats", stats);
     combine_with_exported_module!(&mut lib, "rhai_sci_cumulative", cum_functions);
     combine_with_exported_module!(&mut lib, "rhai_sci_int_and_diff", int_and_diff);
     combine_with_exported_module!(&mut lib, "rhai_sci_assertions", assert_functions);
+    combine_with_exported_module!(&mut lib, "rhai_sci_constants", constant_definitions);
     engine.register_global_module(rhai::Shared::new(lib));
 
     // Extract metadata
@@ -147,7 +148,8 @@ struct Function {
 
 include!("src/matrix.rs");
 include!("src/basic_statistics.rs");
-include!("src/utils.rs");
+include!("src/misc.rs");
 include!("src/cumulative.rs");
 include!("src/integration_and_differentiation.rs");
 include!("src/assertions.rs");
+include!("src/constants.rs");
